refactor: 홈 뷰 컴포넌트 분리
